Separate `AccelEngine` out to header file
